Move row_reader out of internal/
