+++ bigtable/CMakeLists.txt
@@ -101,8 +101,6 @@ add_library(bigtable_client
     client/internal/conjunction.h
     client/internal/readrowsparser.h
     client/internal/readrowsparser.cc
-    client/internal/row_reader.h
-    client/internal/row_reader.cc
     client/filters.h
     client/filters.cc
     client/idempotent_mutation_policy.h
@@ -110,6 +108,8 @@ add_library(bigtable_client
     client/mutations.h
     client/mutations.cc
     client/row.h
+    client/row_reader.h
+    client/row_reader.cc
     client/row_set.h
     client/rpc_backoff_policy.h
     client/rpc_backoff_policy.cc
@@ -174,11 +174,11 @@ set(bigtable_client_unit_tests
     client/idempotent_mutation_policy_test.cc
     client/internal/bulk_mutator_test.cc
     client/internal/readrowsparser_test.cc
-    client/internal/row_reader_test.cc
     client/mutations_test.cc
     client/table_apply_test.cc
     client/table_bulk_apply_test.cc
     client/table_test.cc
+    client/row_reader_test.cc
     client/row_test.cc
     client/rpc_backoff_policy_test.cc
     client/rpc_retry_policy_test.cc)
