+++ bigtable/client/table.h
@@ -18,8 +18,8 @@
 #include "bigtable/client/data_client.h"
 #include "bigtable/client/filters.h"
 #include "bigtable/client/idempotent_mutation_policy.h"
-#include "bigtable/client/internal/row_reader.h"
 #include "bigtable/client/mutations.h"
+#include "bigtable/client/row_reader.h"
 #include "bigtable/client/row_set.h"
 #include "bigtable/client/rpc_backoff_policy.h"
 #include "bigtable/client/rpc_retry_policy.h"
